video: tegra: host: Use kmalloc for jobs up to 8K
Some jobs we now have are more than 4K in size which causes their
allocation to use vzalloc, which is too slow. Increase the threshold
to 8K. In the future we should also decrease the size of job
allocations.

Also print a warning when a job is very large as this really
shouldn't happen.

Bug 200425341
Jira HOSTX-354

Change-Id: I38cc198e69b344af3e6024c13efccf7f27beea82
Signed-off-by: Mikko Perttunen <mperttunen@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1791939
Tested-by: Soumen Dey <sdey@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>