net: wireless: bcmdhd: Move tegra_downgrade_ac to dhd_linux.c
The dhd_custom_sysfs_tegra is a sub-feature of the dhd drivers
to have an sysfs for debugging purpose. So it should not have
a global variable declared which gets extern in the dhd_linux.c
core file -- tegra_downgrade_ac would be missing if the sysfs
config is disabled.

This change just simply exchanges the declare with the extern
in the dhd_linux core file to fix the problem.

Bug 2284925

Change-Id: I9b561f4b3fc7cf335e5d41fe6e2aeac11a428d89
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1800736
GVS: Gerrit_Virtual_Submit
Reviewed-by: Manikanta Maddireddy <mmaddireddy@nvidia.com>
Reviewed-by: Sai Kiran <saik@nvidia.com>
Reviewed-by: Daniel Fu <danifu@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>