driver: therm_fan_est: return trip temp only for PID
When the temp is drop down, there is a chance that the PWM FAN
still get stuck in high speed.

therm_fan_est_get_trip_temp() will retrun the trip temp with hysteresis.
This is to support the hysteresis when the governor didn't support it.
But for PID governor, it will support hysteresis by default.
In such case, it will break the PID throttle(redunantly handle the hysteresis)
and Causing the Fan fail to slow down in above corner case.

Return the trip temp wihout hysteresis, if the governor is PID.

Bug 2256802

Change-Id: Id9d30f207185e33bb9065a29678b85e33e03eadd
Signed-off-by: Daniel Fu <danifu@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1801022
Reviewed-by: Martin Gao <marting@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Vinayak Pane <vpane@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>