video: tegra: dc: support to switch hdmi colorimetry
This change gives support to switch to rec709 or rec2020
colorimentry format on the fly when requested by usermode.
The use case is when native ui is in rec709 and rec2020
contents requested, instead of switching mode, just
switch colorimentry. There are also scenarios to switch to
rec709 from rec2020 so this change also adds support for that.

bug 200352101

Change-Id: If637c34cb45065b55dd522d216b019421f792648
Signed-off-by: Prafull Suryawanshi <prafulls@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1975831
GVS: Gerrit_Virtual_Submit
Reviewed-by: Aly Hirani <ahirani@nvidia.com>
Reviewed-by: Tuomas Kulve <tkulve@nvidia.com>
Reviewed-by: Mitch Luban <mluban@nvidia.com>
Tested-by: Tuomas Kulve <tkulve@nvidia.com>