dp: add Type-C extcon support
In Type-C multi-function pin assignments, the four
USB3.x SS data lanes are divided up for both Alt
Mode and USB signaling. For DP, this means that the
number of available TX lanes is also dependent on the
current pin configuration when driving DP Alt Mode
over Type-C.

The number of connected DP lanes can be derived through
CR/EQ fallback, but this is too late given our current
SW architecture. Mode filtering occurs when the display
driver populates the modedb in response to an HPD_PLUG
event, but LT is only triggered on the subsequent modeset.
In order to resolve this dependency, this change adds
support to the DP driver so that it can query the number
of connected lanes from the ucsi_ccg extcon provider,
before filtering.

Bug 2024637
TDS-3502

Change-Id: I9d0966bc3de5f9323f7ebe18d1fecbe972f6a063
Signed-off-by: Shu Zhong <shuz@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1815326
GVS: Gerrit_Virtual_Submit
Reviewed-by: Mitch Luban <mluban@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>