video: tegra: dc: improve pixel clock accuracy
fbdev framework stores pixelclock in pico seconds. Since monitors
advertize pixclock in hz, we need to convert it from Hz to pico
seconds when creating mode database. During modeset, it again needs
to be converted back to Hz. These conversions cause precision loss,
because of which the final pixclock value does not match the original
value advertized by the sink.

This issue is prominently seen in case of non-CEA modes.

Most of the monitors handle this slight variation in pixclock without
complaining. But, there are a few monitors that fail when the exact
pixclock is not programmed. Hence, interoperability is affected.

Avoiding the pixclock conversion helps in avoiding the precision loss
issue. But, the idea of updating the core kernel fbdev driver to use
pixel clock in hz intead of pico seconds was rejected because it might
cause ABI compatibility issues.

The finalized solution is to store the original pixclock in hz along
with the pixclock in pico seconds, and use the hz pixclock during
modeset instead of the converted one.

Implementation Details:

1. Data structures:

fb_var_screeninfo is the  structure used by userspace clients to specify
their chosen mode timings to kernel. To avoid breaking the current fbdev
ABI, we avoided making any changes to this structure.

A new variable, pixclock_hz, has been added to the existing fb_videomode
structure. It is used to store the original pixclock value in hz during
mode database creation. As part of modeset, this value is used instead of
the one that went through conversion between hz and pico seconds. This
in-turn helps in setting accurate pixel clock advertized by sinks.

2. New defconfig option:

To make it easy to enable/disable this feature, the whole implementation
has been wrapped under a new defconfig option, CONFIG_FB_MODE_PIXCLOCK_HZ.
Any updates to existing data structures has also been isolated under this

3. Fractional Modes:

Fractional modes and their corresponding parent modes are not
differentiated in the fb mode matching API. Since the parent modes are
usually in the beginning of modelist, we always get the non-fractional
mode as matching mode even when fractional modes are requested during
modeset. This had been the case earlier too, and HDMI driver used to
compensate by calculating the fractional pixclock value during modeset.
Similar to this, necessary logic has been added to use fractional
pixelclock if a fractional mode is requested.

4. Removed pclk rounding-off logic in HDMI driver:

tegra_hdmi_get_pclk() was used in HDMI driver to re-calculate pixclock
from mode parameters. This was supposed to help in deriving a more
accurate pclk compared to the Pico Seconds to Hz converted pixclock.
With the support to use the actual pixclock in Hz, this API is no more
needed.

bug 2045478

Change-Id: I33d833b4a8ed2b059568bd595ea73019bc9b0584
Signed-off-by: Naveen Kumar S <nkumars@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1812546
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>