video: tegra: dc: update edid checksum logic
If an EDID is valid, its checksum should be 0. If
not, driver re-reads the EDID a few times to rule
out any intermittent issues that might have corrupted
the EDID.

If current checksum is not 0, this value is saved
in last_checksum variable to compare it with the
checksum in next attempt (if that attempt too fails).
This is done so that we can ascertain if the failed
checksum value is same across multiple read attempts.
In summaray, last_checksum holds the last failed
checksum value.

Hence, while determining if we need to re-read EDID, we
need to check if the current checksum is not 0. Using
last_checksum here is not right as it will always have
the value of failed checksum.

bug 200436277

Change-Id: Ib68d7fd9e87e57f905506c694b359be8f0dd7c00
Signed-off-by: Naveen Kumar S <nkumars@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1810116
Reviewed-by: Shu Zhong <shuz@nvidia.com>
Reviewed-by: Santosh Galma <galmar@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>