hid: jarvis: Add bad ts joystick value stats
There is an issue with Thunderstrike joysticks reporting off-center
when they are not off-center. This change tracks the Thunderstrike
returning bad joystick values at disconnect and exports a table in
format of (mac_address, count) via a driver attribute which can be
utilized by the stats pipeline.

Bug 2462066

Change-Id: I98e984c6350dee4f4b5f9ec320ea444bb1ff27da
Signed-off-by: Andrew Chen <andrewc@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1977780
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Tested-by: Lynette Zhao <lynettez@nvidia.com>
Reviewed-by: Ankita Garg <ankitag@nvidia.com>
Reviewed-by: Vinayak Pane <vpane@nvidia.com>